feat: combine serde types, cleanup unused imports
