+++ packages/types/src/index.ts
@@ -6,11 +6,10 @@ export * from "./crypto";
 export * from "./exception";
 export * from "./http";
 export * from "./logger";
-export * from "./serializer";
+export * from "./serde";
 export * from "./middleware";
 export * from "./protocol";
 export * from "./response";
 export * from "./signature";
-export * from "./deserializer";
 export * from "./transfer";
 export * from "./util";
